\name{A2Rplot}
\alias{A2Rplot}
\alias{A2Rplot.default}
\title{Generic plot function for package A2R}
\description{That function is a generic function for customised 
  plotting. Currently, only A2Rplot.hclust is on.}
\arguments{
  \item{x}{main argument}
  \item{...}{others parameters to pass to specific methods}
}
\seealso{
  \code{\link{A2Rplot.hclust}}
}
\keyword{hplot}

